Created small modules
Note that:
demux_1_4 is a 1-4 demux
jk_ff is a jk flip flop, with an asynchronous CLR signal which automatically sets the output to 0 regardless of the state of the clock. It's how the paper described their flipflops, not sure if it's important that it be async. However my code doesn't test well - just outputs x no matter what.
jk_ff_sync is a jk flip flop but makes the CLR synchronous. Also doesn't work